Rewriten with lombok
