--- src/main/java/Radio.java
@@ -1,75 +1,0 @@
-public class Radio {
-    private int currentStation;
-    private int currentVolume;
-
-    private int numberStations;
-
-    public Radio() {
-        this.numberStations = 10;
-    }
-
-    public Radio(int numberStations) {
-        this.numberStations = numberStations;
-    }
-
-    public void next() {
-        if (currentStation < numberStations - 1) {
-            currentStation++;
-        } else {
-            currentStation = 0;
-        }
-    }
-
-    public void prev() {
-        if (currentStation > 0) {
-            currentStation--;
-        } else {
-            currentStation = numberStations - 1;
-        }
-    }
-
-    public void increaseVolume() {
-        if (currentVolume < 100) {
-            currentVolume++;
-        } else {
-            currentVolume = 100;
-        }
-    }
-
-    public void decreaseVolume() {
-        if (currentVolume > 0) {
-            currentVolume--;
-        } else {
-            currentVolume = 0;
-        }
-    }
-
-    public int getCurrentStation() {
-        return currentStation;
-    }
-
-    public void setCurrentStation(int currentStation) {
-        if (currentStation < 0) {
-            return;
-        }
-        if (currentStation > numberStations - 1) {
-            return;
-        }
-        this.currentStation = currentStation;
-    }
-
-    public int getCurrentVolume() {
-        return currentVolume;
-    }
-
-    public void setCurrentVolume(int currentVolume) {
-        if (currentVolume < 0) {
-            return;
-        }
-        if (currentVolume > 100) {
-            return;
-        }
-        this.currentVolume = currentVolume;
-    }
-}
-
