+++ src/test/java/ru/netology/radio/RadioTest.java
@@ -1,3 +1,5 @@
+package ru.netology.radio;
+
 import static org.junit.jupiter.api.Assertions.*;
 import org.junit.jupiter.api.Test;
 
@@ -173,5 +175,4 @@ public class RadioTest {
         Radio radio = new Radio(30);
         this.numberStations = numberStations;
     }
-
 }
